backend/pyproject.toml
[project]
name = "tbo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.57",
    "python-dotenv>=1.0.1",
    "fastapi>=0.133.1",
    "google-genai>=1.65.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-openai>=0.3.28",
    "langgraph>=1.0.9",
    "openai>=2.24.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.3",
    "uvicorn>=0.41.0",
    "botocore>=1.42.58",
    "python-multipart>=0.0.22",
]
